+++ cranelift/codegen/shared/src/lib.rs
@@ -19,7 +19,6 @@
     )
 )]
 
-pub mod condcodes;
 pub mod constant_hash;
 pub mod constants;
